+++ tests/unit/libshvcp/cpdecimal.c
@@ -9,16 +9,16 @@ TEST_CASE(all) {}
 static const struct {
 	struct cpdecimal val, norm;
 } decnorm_d[] = {
-	{{.mantisa = 1}, {.mantisa = 1}},
-	{{.mantisa = 100}, {.mantisa = 1, .exponent = 2}},
-	{{.mantisa = 100, .exponent = -7}, {.mantisa = 1, .exponent = -5}},
+	{{.mantissa = 1}, {.mantissa = 1}},
+	{{.mantissa = 100}, {.mantissa = 1, .exponent = 2}},
+	{{.mantissa = 100, .exponent = -7}, {.mantissa = 1, .exponent = -5}},
 	{{}, {}},
 	{{.exponent = 100}, {}},
 };
 ARRAY_TEST(all, decnorm) {
 	struct cpdecimal d = _d.val;
 	cpdecnorm(&d);
-	ck_assert_int_eq(d.mantisa, _d.norm.mantisa);
+	ck_assert_int_eq(d.mantissa, _d.norm.mantissa);
 	ck_assert_int_eq(d.exponent, _d.norm.exponent);
 }
 END_TEST
@@ -28,10 +28,10 @@ static const struct {
 	struct cpdecimal d;
 } dec_d[] = {
 	{0., (struct cpdecimal){}},
-	{1., (struct cpdecimal){.mantisa = 1}},
-	{-1., (struct cpdecimal){.mantisa = -1}},
-	{42.124, (struct cpdecimal){.mantisa = 42124, .exponent = -3}},
-	{400., (struct cpdecimal){.mantisa = 4, .exponent = 2}},
+	{1., (struct cpdecimal){.mantissa = 1}},
+	{-1., (struct cpdecimal){.mantissa = -1}},
+	{42.124, (struct cpdecimal){.mantissa = 42124, .exponent = -3}},
+	{400., (struct cpdecimal){.mantissa = 4, .exponent = 2}},
 };
 
 ARRAY_TEST(all, dectod, dec_d) {
@@ -54,7 +54,7 @@ static const struct {
 ARRAY_TEST(all, decexp) {
 	struct cpdecimal d = _d.in;
 	ck_assert(cpdecexp(&d, _d.exponent));
-	ck_assert_int_eq(d.mantisa, _d.out.mantisa);
+	ck_assert_int_eq(d.mantissa, _d.out.mantissa);
 	ck_assert_int_eq(d.exponent, _d.out.exponent);
 }
 
@@ -84,14 +84,14 @@ ARRAY_TEST(all, cpdtoi) {
 }
 
 static const struct {
-	int mantisa, exponent;
+	int mantissa, exponent;
 	struct cpdecimal dec;
 } cpitod_d[] = {
 	{42, -3, {42, -3}},
 	{4200, -3, {42, -1}},
 };
 ARRAY_TEST(all, cpitod) {
-	struct cpdecimal d = cpitod(_d.mantisa, _d.exponent);
-	ck_assert_int_eq(d.mantisa, _d.dec.mantisa);
+	struct cpdecimal d = cpitod(_d.mantissa, _d.exponent);
+	ck_assert_int_eq(d.mantissa, _d.dec.mantissa);
 	ck_assert_int_eq(d.exponent, _d.dec.exponent);
 }